newsapi-drf-levelone/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
flake8>=4.0.1<4.1.0
